% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matrix.R
\name{GeraMatrizHilbert}
\alias{GeraMatrizHilbert}
\title{Gera matriz de Hilbert}
\usage{
GeraMatrizHilbert(dimensao)
}
\arguments{
\item{dimensao}{dimensão da matriz a ser gerada (nxn)}
}
\value{
Retorna a matriz de Hilbert, de tamanhao nxn (quadrada)
}
\description{
Gera uma matriz de Hilbert, onde cada termo a_ij = 1 / (i + j - 1)
}
