% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_solvers.R
\name{solver_comPivotamento}
\alias{solver_comPivotamento}
\title{Solver com pivotamento}
\usage{
solver_comPivotamento(matriz, vetor)
}
\arguments{
\item{matriz}{Matriz quadrada, de ordem n}

\item{vetor}{Vetor de dimensao nx1}
}
\value{
Retorna o vetor solução 'x' e o determinante da matriz
}
\description{
Recebe uma matriz e um vetor 'b' e realiza a eliminação de Gauss com pivotamento.
}
